Do as much as possible at compile time
Before we cached the functions on first run and funcall'd them, now we
just use a macro to expand to the actual function call.

For the connection we used to look it up and cache it, now we access
the variable (or call the fallback function). Not sure if this is
slower..but will respond better to a new connection being made